% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epict-package.R
\docType{package}
\name{epict-package}
\alias{epict}
\alias{epict-package}
\title{epict: Individual level Cycle Threshold Modelling}
\description{
Individual level Cycle Threshold Modelling
}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Tim Russell \email{timothy.russell@lshtm.ac.uk}
  \item Joel Hellewell \email{joel.hellewell@lshtm.ac.uk}
}

}
\keyword{internal}
