% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_check_raw_obs}
\alias{epict_check_raw_obs}
\title{Check unproccessed observations meet the package specification}
\usage{
epict_check_raw_obs(obs, dates_available = TRUE, check_onset = FALSE)
}
\arguments{
\item{obs}{A data.frame with the following variables:
\itemize{
\item id: An integer vector uniquely identifying eahc infection.
\item test_id: An integer vector uniquely identiying each test
\item ct_value: Numeric cycle threshold value.
\item test_date: Date of the test yielding a Ct value. (optional)
\item t: Time of test relative to a baseline date. Optional but required
if test_date is not present.
\item onset_date Date of onset for each infection (optional).
NA if unavailable/asymptomatic
\item onset_t Time on onset relative to a baseline date (optional).
\item censored: Logical, indicating if the Ct has been censored.
}}

\item{dates_available}{Logical, defaults to \code{TRUE}. Are dates available
in the observed data. If \code{FALSE} it is assumed that relative times should be
available instead.}

\item{check_onset}{Logical, defaults to \code{FALSE}. Should observations be
checked for the presence of symptom onset data.}
}
\value{
Input observations are returned.
}
\description{
Checks the available raw data for required variables and returns informative
errors, warnings, and messages about the structure of the observations. It
can be used both on datasets referenced using dates and datasets referenced
using relative time.
}
\author{
Sam Abbott
}
