% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_clean_factors}
\alias{epict_clean_factors}
\title{Clean up factor levels}
\usage{
epict_clean_factors(vars = c())
}
\arguments{
\item{vars}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Sam Abbott
}
