% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{epict_convert_to_list}
\alias{epict_convert_to_list}
\title{Format observations and model settings for use with stan}
\usage{
epict_convert_to_list(
  obs,
  piecewise_formula = epict::piecewise_formula(~1, obs),
  adjustment_formula = epict::adjustment_formula(~1, obs),
  priors = epict::epict_priors(),
  model_opts = epict::epict_model_opts(),
  inference_opts = epict::epict_inference_opts()
)
}
\arguments{
\item{obs}{A data.frame with the following variables:
\itemize{
\item \code{id}: An integer vector uniquely identifying eahc infection.
\item \code{test_id}: An integer vector uniquely identiying each test
\item \code{ct_value}: Numeric cycle threshold value.
\item \code{t}: Relative (to a baseline) time of the test yielding a Ct value.
\item \code{t_rel_uncensored}: Time of test relative to the first uncensored Ct value
for that id.
\item \code{onset_t}: Relative (to a baseline) time of onset for each infection
\item \code{onset_t_rel_uncensored}: Time of onset relative to the first uncensored
Ct value for that id. (optional). NA if unavailable/asymptomatic.
\item \code{censored}: Logical, indicating if the Ct has been censored.
}}

\item{piecewise_formula}{A list describing the piecewise linear cycle threshold
formula as described by \code{\link[=piecewise_formula]{piecewise_formula()}}.}

\item{adjustment_formula}{A list describing the cycle threshold linear
adjustment formula (shift and scale) as described by \code{\link[=adjustment_formula]{adjustment_formula()}}.}

\item{priors}{A data.table of population-level priors as produced
by \code{\link[=epict_priors]{epict_priors()}}.}

\item{model_opts}{A list of model options. See \code{\link[=epict_model_opts]{epict_model_opts()}}
for details}

\item{inference_opts}{A list of options to use for inference. See
\code{\link[=epict_inference_opts]{epict_inference_opts()}} for details.}
}
\value{
A list as required by stan.
}
\description{
Format observations and model settings for use with stan
}
\seealso{
Functions used for modelling
\code{\link{epict_inits}()},
\code{\link{epict_model}()},
\code{\link{epict_priors}()}
}
\author{
Sam Abbott
}
\concept{model}
