% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{epict_drop_na_ct}
\alias{epict_drop_na_ct}
\title{Drop Ct values with no data}
\usage{
epict_drop_na_ct(obs)
}
\arguments{
\item{obs}{A \code{data.frame} with a numeric \code{ct_value} variable.}
}
\value{
The input \code{data.table} with \code{NA} values dropped.
}
\description{
Drop Ct values with no data
}
\examples{
obs <- data.frame(ct_value = c(1, NA, 2))
epict_drop_na_ct(obs)
}
\seealso{
Preprocessing functions
\code{\link{epict_check_obs}()},
\code{\link{epict_check_raw_obs}()},
\code{\link{epict_clean_factors}()},
\code{\link{epict_filter_ids}()},
\code{\link{epict_flag_spurious_obs}()},
\code{\link{epict_make_time_rel_to_first_uncensored}()},
\code{\link{epict_make_time_rel}()}
}
\author{
Sam Abbott
}
\concept{preprocess}
