% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_filter_ids}
\alias{epict_filter_ids}
\title{Filter infection IDs based on characteristics}
\usage{
epict_filter_ids(
  obs,
  min_uncensored_tests = 2,
  min_days_with_uncensored = 2,
  invert = FALSE
)
}
\arguments{
\item{obs}{DESCRIPTION.}

\item{min_uncensored_tests}{DESCRIPTION.}

\item{min_days_with_uncensored}{DESCRIPTION.}

\item{invert}{Logical, defaults to \code{FALSE}. Should the filtering
requirements be inverted.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Sam Abbott
}
