% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_flag_spurious_ids}
\alias{epict_flag_spurious_ids}
\title{Flag potentially spurious IDs}
\usage{
epict_flag_spurious_ids(
  max_t_rel_uncensored = 60,
  max_onset_t_rel_uncensored = 15,
  flag = TRUE,
  drop = TRUE,
  return_spurious = FALSE
)
}
\arguments{
\item{max_t_rel_uncensored}{DESCRIPTION.}

\item{max_onset_t_rel_uncensored}{DESCRIPTION.}

\item{flag}{DESCRIPTION.}

\item{drop}{DESCRIPTION.}

\item{return_spurious}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Sam Abbott
}
