% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{epict_flag_spurious_obs}
\alias{epict_flag_spurious_obs}
\title{Flag potentially spurious observations}
\usage{
epict_flag_spurious_obs(
  obs,
  max_t_rel_uncensored = 60,
  max_onset_t_rel_uncensored = 15,
  flag = TRUE,
  drop = TRUE,
  return_spurious = FALSE
)
}
\arguments{
\item{obs}{A data.frame with the following variables:
\itemize{
\item \code{t_rel_uncensored}: Time of test relative to the first uncensored Ct value
for that id.
\item \code{onset_t_rel_uncensored}: Time of onset relative to the first uncensored
Ct value for that id. (optional). NA if unavailable/asymptomatic.
}}

\item{max_t_rel_uncensored}{Numeric defaults to 60 days. Flags the maximum
absolute relative time that is considered plausible for tests to be
spaced by per ID.}

\item{max_onset_t_rel_uncensored}{Numeric defaults to 15 days. Flags the
maximum absolute relative time that is considered plausible for symptom
onset from the first uncensored test per ID.}

\item{flag}{Logical, defaults to \code{TRUE}. Should spurious tests be flagged}

\item{drop}{Logical, defaults to \code{TRUE}. Should spurious tests
and onsets be dropped (set to \code{NA} for onsets and filtered out for tests).}

\item{return_spurious}{Logical, defaults to \code{FALSE}. Rather than returning
observations should observations flagged as spurious be returned.}
}
\value{
A \code{data.table} of observations
}
\description{
Flag potentially spurious observations
}
\examples{
obs <- data.frame(t_rel_uncensored = c(0, 2, 10, 60, 100, 30))

# Run with defaults
fil_obs <- epict_flag_spurious_obs(obs)
fil_obs
epict_flag_spurious_obs(fil_obs)

# Add onsets and repeat
obs$onset_t_rel_uncensored <- c(0, 15, 2, NA, 40, 5)
fil_obs <- epict_flag_spurious_obs(obs)
fil_obs
epict_flag_spurious_obs(fil_obs)

# Return spurious observations
epict_flag_spurious_obs(obs, return_spurious = TRUE)

# Flag spurious results but don't drop
epict_flag_spurious_obs(obs, drop = FALSE)

# Siltently drop
epict_flag_spurious_obs(obs, flag = FALSE)
}
\author{
Sam Abbott
}
