% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{epict_inits}
\alias{epict_inits}
\title{Set up initial conditions for model}
\usage{
epict_inits(data, scale = 0.1)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=epict_as_list]{epict_as_list()}}.}

\item{scale}{Numeric defaults to 0.1. The numeric scaling to apply to the
standard deviation when sampling. Helpful for reducing numeric instability
during warmup.}
}
\value{
A function that when called returns a list of initial conditions
for the package stan models.
}
\description{
Set up initial conditions for model
}
\seealso{
Functions used for modelling
\code{\link{epict_convert_to_list}()},
\code{\link{epict_model}()},
\code{\link{epict_priors}()}
}
\author{
Sam Abbott
}
\concept{model}
