% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_make_time_rel}
\alias{epict_make_time_rel}
\title{Make time relative to first test in the data}
\usage{
epict_make_time_rel(obs, baseline_date = min(obs$onset_date, na.rm = TRUE))
}
\arguments{
\item{obs}{DESCRIPTION.}

\item{baseline_date}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Sam Abbott
}
