% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_make_time_rel_to_first_uncensored}
\alias{epict_make_time_rel_to_first_uncensored}
\title{Make time relative to first uncensored test per ID}
\usage{
epict_make_time_rel_to_first_uncensored(obs)
}
\arguments{
\item{obs}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Sam Abbott
}
