% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_model_opts}
\alias{epict_model_opts}
\title{Format formula data for use with stan}
\usage{
epict_model_opts(
  onsets = TRUE,
  switch = FALSE,
  latent_infections = TRUE,
  variation = "correlated"
)
}
\arguments{
\item{switch}{Logical, default to \code{FALSE}. Should a secondary breakpoint be
included in the piecewise linear Cycle threshold model.}

\item{latent_infections}{Logical, defaults to \code{TRUE}. Should latent time of
infection be modelled and used to adjust time from first positive test.}

\item{variation}{A character string indicating the type of individual level
variation to include. Defaults to "correlated" (a random effect with
modelled correlation structure). Other options include "uncorrelated" (
a random effect with no modelled correlation structure), and "none" (for no
individual level variation).}
}
\value{
A list as required by stan.
}
\description{
Format formula data for use with stan
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{epict_formula_as_list}()},
\code{\link{epict_individual_priors_as_list}()},
\code{\link{epict_inference_opts}()},
\code{\link{epict_obs_as_list}()},
\code{\link{epict_onset_obs_as_list}()},
\code{\link{epict_population_priors_as_list}()},
\code{\link{epict_posterior_as_prior}()}
}
\author{
Sam Abbott
}
\concept{modeltools}
