% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_obs_as_list}
\alias{epict_obs_as_list}
\title{Format observed data for use with stan}
\usage{
epict_obs_as_list(obs)
}
\arguments{
\item{obs}{DESCRIPTION.}
}
\value{
A list as required by stan.
}
\description{
Formats observations for use in the stan model. Note that the
censoring limit is set internally in this function based on the minimum
cycle threshold value present for a censored test.
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{epict_formula_as_list}()},
\code{\link{epict_individual_priors_as_list}()},
\code{\link{epict_inference_opts}()},
\code{\link{epict_model_opts}()},
\code{\link{epict_onset_obs_as_list}()},
\code{\link{epict_population_priors_as_list}()},
\code{\link{epict_posterior_as_prior}()}
}
\author{
Sam Abbott
}
\concept{modeltools}
