% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_onset_obs_as_list}
\alias{epict_onset_obs_as_list}
\title{Format observed symptom onset data for use with stan}
\usage{
epict_onset_obs_as_list(obs, onsets = TRUE)
}
\arguments{
\item{obs}{DESCRIPTION.}

\item{onsets}{Logical, defaults to \code{TRUE}. Should symptom onsets
observations be included in the model if available.}
}
\value{
A list as required by stan.
}
\description{
Format observed symptom onset data for use with stan
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{epict_formula_as_list}()},
\code{\link{epict_individual_priors_as_list}()},
\code{\link{epict_inference_opts}()},
\code{\link{epict_model_opts}()},
\code{\link{epict_obs_as_list}()},
\code{\link{epict_population_priors_as_list}()},
\code{\link{epict_posterior_as_prior}()}
}
\author{
Sam Abbott
}
\concept{modeltools}
