% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_opts_as_data_list}
\alias{epict_opts_as_data_list}
\title{Format model options for use with stan}
\usage{
epict_opts_as_data_list(
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE
)
}
\arguments{
\item{pp}{Logical, defaults to \code{FALSE}. Should posterior predictions be made
for observed data. Useful for evaluating the performance of the model.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}
}
\value{
A list as required by stan.
}
\description{
Format model options for use with stan
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{epict_formula_as_data_list}()},
\code{\link{epict_individual_priors_as_data_list}()},
\code{\link{epict_population_priors_as_data_list}()},
\code{\link{epict_posterior_as_prior}()}
}
\concept{modeltools}
