% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_population_priors_as_data_list}
\alias{epict_population_priors_as_data_list}
\title{FUNCTION_TITLE}
\usage{
epict_population_priors_as_data_list(priors)
}
\arguments{
\item{priors}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
priors <- epict_priors()
epict_population_priors_as_data_list(priors)
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{epict_formula_as_data_list}()},
\code{\link{epict_individual_priors_as_data_list}()},
\code{\link{epict_opts_as_data_list}()},
\code{\link{epict_posterior_as_prior}()}
}
\concept{modeltools}
