% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_posterior_as_prior}
\alias{epict_posterior_as_prior}
\title{Extract summarised posteriors to use as priors}
\usage{
epict_posterior_as_prior(
  fit,
  priors = epict::epict_priors(),
  variables = c(),
  scale = 5
)
}
\arguments{
\item{fit}{A \link{cmdstanr} fit as returned by \code{\link[=epict]{epict()}}.}

\item{priors}{A data frame of priors to update as defined in
\code{\link[=epict_priors]{epict_priors()}}.}

\item{variables}{A character vector of variables both in the
posterior and in the default priors.}

\item{scale}{Numeric, defaults to 5. Amount to scale posterior standard
deviations by.}
}
\value{
A data frame of priors
}
\description{
Extract and summarise posteriors from a \code{\link[=epict]{epict()}} to use as
priors (based on the distributions given in \code{\link[=epict_priors]{epict_priors()}})
assuming they are independent normal
}
\seealso{
Other modeltools: 
\code{\link{epict_formula_as_data_list}()},
\code{\link{epict_opts_as_data_list}()},
\code{\link{epict_priors_as_data_list}()}
}
\concept{modeltools}
