% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{epict_priors}
\alias{epict_priors}
\title{Default model priors}
\usage{
epict_priors()
}
\value{
A data frame summarising the model priors.
}
\description{
Default model priors
}
\examples{
epict_priors()
}
\seealso{
Other model: 
\code{\link{epict_inits}()},
\code{\link{epict_model}()},
\code{\link{epict_to_stan}()}
}
\concept{model}
