% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{epict_to_stan}
\alias{epict_to_stan}
\title{Format data for use with stan}
\usage{
epict_to_stan(
  obs,
  ct_model = subject_design(~1, obs),
  adjustment_model = test_design(~1, obs),
  priors = epict::epict_priors(),
  individual_variation = 0.2,
  individual_correlation = 1,
  censoring_threshold = 40,
  switch = TRUE,
  onsets = TRUE,
  incubation_period = get_inc_period(),
  likelihood = TRUE,
  output_loglik = FALSE
)
}
\value{
A list as required by stan.
}
\description{
Format data for use with stan
}
\seealso{
Functions used for modelling
\code{\link{epict_inits}()},
\code{\link{epict_model}()},
\code{\link{epict_priors}()}
}
\concept{model}
