% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{epict_update_variable_labels}
\alias{epict_update_variable_labels}
\title{FUNCTION_TITLE}
\usage{
epict_update_variable_labels(
  draws,
  reverse = FALSE,
  variables = epict::epict_priors(include_descriptive = TRUE)
)
}
\arguments{
\item{draws}{DESCRIPTION.}

\item{reverse}{DESCRIPTION.}

\item{variables}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
\author{
Sam Abbott
}
\concept{postprocess}
