% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{adjustment_formula}
\alias{adjustment_formula}
\title{Specify the cycle threshold adjustment model formula}
\usage{
adjustment_formula(formula = ~1, obs, beta_default = c(0, 0.1))
}
\arguments{
\item{formula}{A model formula defaulting to \code{~1}.}

\item{obs}{A data.frame with the following variables:
\itemize{
\item \code{id}: An integer vector uniquely identifying eahc infection.
\item \code{test_id}: An integer vector uniquely identiying each test
\item \code{ct_value}: Numeric cycle threshold value.
\item \code{t}: Relative (to a baseline) time of the test yielding a Ct value.
\item \code{t_rel_uncensored}: Time of test relative to the first uncensored Ct value
for that id.
\item \code{onset_t}: Relative (to a baseline) time of onset for each infection
\item \code{onset_t_rel_uncensored}: Time of onset relative to the first uncensored
Ct value for that id. (optional). NA if unavailable/asymptomatic.
\item \code{censored}: Logical, indicating if the Ct has been censored.
}}

\item{beta_default}{A vector of length two containing the
default mean and standard deviation to use as the prior
for all covariate effect sizes.}
}
\value{
A named list including the design matrix ("design")
and a \code{data.table} of priors for covariate effects ("beta").
}
\description{
Specify the cycle threshold adjustment model formula
}
\author{
Sam Abbott
}
