% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{adjustment_formula}
\alias{adjustment_formula}
\title{Specify the cycle threshold adjustment model formula}
\usage{
adjustment_formula(formula = ~1, obs, beta_default = c(0, 0.1))
}
\arguments{
\item{formula}{A model formula defaulting to \code{~1}.}

\item{obs}{DESCRIPTION.}

\item{beta_default}{A vector of length two containing the
default mean and standard deviation to use as the prior
for all covariate effect sizes.}
}
\value{
A named list including the design matrix ("design")
and a \code{data.table} of priors for covariate effects ("beta").
}
\description{
Specify the cycle threshold adjustment model formula
}
\author{
Sam Abbott
}
