% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epict-package.R
\docType{package}
\name{epict-package}
\alias{epict}
\alias{epict-package}
\title{epict: Flexible Cycle Threshold Modelling}
\description{
Flexible piecewise modelling of cycle thresholds with adjustment for varying intercepts and linear scales.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://samabbott.co.uk/epict/}
  \item \url{https://github.com/seabbs/epict/}
  \item Report bugs at \url{https://github.com/seabbs/epict/issues/}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Tim Russell \email{timothy.russell@lshtm.ac.uk}
  \item Joel Hellewell \email{joel.hellewell@lshtm.ac.uk}
}

}
\keyword{internal}
