% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epict.R
\name{epict}
\alias{epict}
\title{Model Cycle Thresholds}
\usage{
epict(
  obs,
  model = epict_model(),
  as_data_list = epict_to_stan,
  inits = epict_inits,
  ct_model = subject_design(~1, obs),
  adjustment_model = test_design(~1, obs),
  individual_variation = 0.2,
  individual_correlation = 1,
  censoring_threshold = 40,
  switch = TRUE,
  onsets = TRUE,
  incubation_period = get_inc_period(),
  likelihood = TRUE,
  output_loglik = FALSE,
  ...
)
}
\arguments{
\item{as_data_list}{PARAM_DESCRIPTION}

\item{inits}{PARAM DESCRIPTION}

\item{...}{Additional arguments passed to \link{cmdstanr}.}
}
\value{
A \link{cmdstanr} model fit containing posterior samples
}
\description{
Provides a user friendly interface around package functionality
to model cycle tresholds, and symptom onsets from observed preprocessed
data, a piecewise linear model, and a global adjustment model.
}
\concept{epict}
