% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{epict_check_obs}
\alias{epict_check_obs}
\title{Check proccessed observations meet the package specification}
\usage{
epict_check_obs(obs, check_onsets = FALSE)
}
\arguments{
\item{obs}{A data.frame with the following variables:
\itemize{
\item \code{id}: An integer vector uniquely identifying eahc infection.
\item \code{test_id}: An integer vector uniquely identiying each test
\item \code{ct_value}: Numeric cycle threshold value.
\item \code{t}: Relative (to a baseline) time of the test yielding a Ct value.
\item \code{t_rel_uncensored}: Time of test relative to the first uncensored Ct value
for that id.
\item \code{onset_t}: Relative (to a baseline) time of onset for each infection
\item \code{onset_t_rel_uncensored}: Time of onset relative to the first uncensored
Ct value for that id. (optional). NA if unavailable/asymptomatic.
\item \code{censored}: Logical, indicating if the Ct has been censored.
}}
}
\value{
A \code{data.table} of observations ready for use in \code{\link[=epict]{epict()}} and other
package functions.
}
\description{
Checks the available proccessed observastionsfor required variables and
returns informative errors, warnings, and messages about the structure of
the observations.
}
\author{
Sam Abbott
}
