% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_check_obs}
\alias{epict_check_obs}
\title{Check processed observations meet package specifications}
\usage{
epict_check_obs(obs, check_onsets = FALSE)
}
\arguments{
\item{obs}{DESCRIPTION.}

\item{check_onsets}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Required processed data format:
\itemize{
\item id: An integer vector uniquely identifying eahc infection.
\item test_id: An integer vector uniquely identiying each test
\item ct_value: Numeric cycle threshold value.
\item t: Relative (to a baseline) time of the test yielding a Ct value.
\item t_rel_uncensored: Time of test relative to the first uncensored Ct value
for that id.
\item onset_t: Relative (to a baseline) time of onset for each infection
\item onset_t_rel_uncensored: Time of onset relative to the first uncensored Ct
value for that id. (optional). NA if unavailable/asymptomatic.
\item censored: Logical, indicating if the Ct has been censored.
}
}
\author{
Sam Abbott
}
