% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{epict_clean_factors}
\alias{epict_clean_factors}
\title{Clean up factor levels}
\usage{
epict_clean_factors(obs, vars = c())
}
\arguments{
\item{obs}{A \code{data.frame} with at least one character or factor variable.}

\item{vars}{A character vector of variables. Defaults to empty.}
}
\value{
A \code{data.table} with empty factors dropped and character vectors
transformed to factors as specified.
}
\description{
This utility function drops empty factor levels from target factors as well
as converting them to factors from character variables.
}
\examples{
obs <- data.frame(
 m = c("fa", "aefwe", "efe"),
 c = factor(c("fa", "asas", "asa"), levels = c("fa"))
)
summary(obs)

# Default without specifying variables
summary(epict_clean_factors(obs))

# Specify variables
summary(epict_clean_factors(obs, vars = c("m", "c")))
}
\seealso{
Preprocessing functions
\code{\link{epict_check_obs}()},
\code{\link{epict_check_raw_obs}()},
\code{\link{epict_drop_na_ct}()},
\code{\link{epict_filter_ids}()},
\code{\link{epict_flag_spurious_obs}()},
\code{\link{epict_make_time_rel_to_first_uncensored}()},
\code{\link{epict_make_time_rel}()}
}
\author{
Sam Abbott
}
\concept{preprocess}
