% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_drop_na_ct}
\alias{epict_drop_na_ct}
\title{Drop Ct values with no data}
\usage{
epict_drop_na_ct(obs)
}
\arguments{
\item{obs}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Sam Abbott
}
