% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-obs.R
\name{epict_drop_na_ct}
\alias{epict_drop_na_ct}
\title{Drop Ct values with no data}
\usage{
epict_drop_na_ct(obs)
}
\arguments{
\item{obs}{A \code{data.frame} with a numeric \code{ct_value} variable.}
}
\value{
The input \code{data.table} with \code{NA} values dropped.
}
\description{
Drop Ct values with no data
}
\examples{
obs <- data.frame(ct_value = c(1, NA, 2))
epict_drop_na_ct(obs)
}
\author{
Sam Abbott
}
