% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{epict_inits}
\alias{epict_inits}
\title{Set up initial conditions for model}
\usage{
epict_inits(data)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=epict_to_stan]{epict_to_stan()}}.}
}
\value{
A function that when called returns a list of initial conditions
for the package stan models.
}
\description{
Set up initial conditions for model
}
\seealso{
Other model: 
\code{\link{epict_model}()},
\code{\link{epict_priors}()},
\code{\link{epict_to_stan}()}
}
\concept{model}
