% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_onset_obs_as_list}
\alias{epict_onset_obs_as_list}
\title{Format observed symptom onset data for use with stan}
\usage{
epict_onset_obs_as_list(obs, onsets = TRUE)
}
\arguments{
\item{obs}{A data.frame with the following variables:
\itemize{
\item \code{id}: An integer vector uniquely identifying eahc infection.
\item \code{test_id}: An integer vector uniquely identiying each test
\item \code{ct_value}: Numeric cycle threshold value.
\item \code{t}: Relative (to a baseline) time of the test yielding a Ct value.
\item \code{t_rel_uncensored}: Time of test relative to the first uncensored Ct value
for that id.
\item \code{onset_t}: Relative (to a baseline) time of onset for each infection
\item \code{onset_t_rel_uncensored}: Time of onset relative to the first uncensored
Ct value for that id. (optional). NA if unavailable/asymptomatic.
\item \code{censored}: Logical, indicating if the Ct has been censored.
}}

\item{onsets}{Logical, defaults to \code{TRUE}. Should symptom onsets
observations be included in the model if available.}
}
\value{
A list as required by stan.
}
\description{
The package stan code (accessed using \code{\link[=epict_model]{epict_model()}}) has
more extensive documentation.
}
\seealso{
Functions used to design and setup models
\code{\link{adjustment_formula}()},
\code{\link{epict_formula_as_list}()},
\code{\link{epict_individual_priors_as_list}()},
\code{\link{epict_inference_opts}()},
\code{\link{epict_model_opts}()},
\code{\link{epict_obs_as_list}()},
\code{\link{epict_population_priors_as_list}()},
\code{\link{epict_posterior_as_prior}()},
\code{\link{piecewise_formula}()},
\code{\link{select_piecewise_parameters}()}
}
\author{
Sam Abbott
}
\concept{modeltools}
