% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{epict_posterior_as_prior}
\alias{epict_posterior_as_prior}
\title{Extract summarised posteriors to use as priors}
\usage{
epict_posterior_as_prior(
  fit,
  priors = epict::epict_priors(),
  variables = c(),
  sub = "_int",
  scale = 5
)
}
\arguments{
\item{fit}{A \link{cmdstanr} fit as returned by \code{\link[=epict]{epict()}}.}

\item{priors}{A data frame of priors to update as defined in
\code{\link[=epict_priors]{epict_priors()}}.}

\item{variables}{A character vector of variables both in the
posterior and in the default priors.}

\item{sub}{A character string indicating the part of variable names in the
posterior to remove. This can be used to link variable names in the model
object to those expected as input.}

\item{scale}{Numeric, defaults to 5. Amount to scale posterior standard
deviations by.}
}
\value{
A data frame of priors
}
\description{
Extract and summarise posteriors from a \code{\link[=epict]{epict()}} to use as
priors (based on the distributions given in \code{\link[=epict_priors]{epict_priors()}})
assuming they are independent normal
}
\seealso{
Functions used to design and setup models
\code{\link{adjustment_formula}()},
\code{\link{epict_formula_as_list}()},
\code{\link{epict_individual_priors_as_list}()},
\code{\link{epict_inference_opts}()},
\code{\link{epict_model_opts}()},
\code{\link{epict_obs_as_list}()},
\code{\link{epict_onset_obs_as_list}()},
\code{\link{epict_population_priors_as_list}()},
\code{\link{piecewise_formula}()},
\code{\link{select_piecewise_parameters}()}
}
\author{
Sam Abbott
}
\concept{modeltools}
