% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{epict_priors}
\alias{epict_priors}
\title{Default model priors}
\usage{
epict_priors(individual_variation = c(0, 0.05))
}
\arguments{
\item{individual_variation}{A vector of length two specifying the
default mean and standard deviation of individual-level variation.}
}
\value{
A data frame summarising the model priors.
}
\description{
Default model priors
}
\examples{
epict_priors()
}
\seealso{
Functions used for modelling
\code{\link{epict_convert_to_list}()},
\code{\link{epict_inits}()},
\code{\link{epict_model}()}
}
\author{
Sam Abbott
}
\concept{model}
