% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{piecewise_formula}
\alias{piecewise_formula}
\title{Specify the cycle threshold adjustment model formula}
\usage{
piecewise_formula(formula = ~1, obs, beta_default = c(0, 0.1), params = "all")
}
\arguments{
\item{formula}{A model formula defaulting to \code{~1}.}

\item{obs}{A data.frame with the following variables:
\itemize{
\item \code{id}: An integer vector uniquely identifying eahc infection.
\item \code{test_id}: An integer vector uniquely identiying each test
\item \code{ct_value}: Numeric cycle threshold value.
\item \code{t}: Relative (to a baseline) time of the test yielding a Ct value.
\item \code{t_rel_uncensored}: Time of test relative to the first uncensored Ct value
for that id.
\item \code{onset_t}: Relative (to a baseline) time of onset for each infection
\item \code{onset_t_rel_uncensored}: Time of onset relative to the first uncensored
Ct value for that id. (optional). NA if unavailable/asymptomatic.
\item \code{censored}: Logical, indicating if the Ct has been censored.
}}

\item{beta_default}{A vector of length two containing the
default mean and standard deviation to use as the prior
for all covariate effect sizes.}

\item{params}{A character string indicating the paramters
to adjust in the piecewise model. Defaults to "all". Options
are: the time at peak ("t_p"), time at switch ("t_p"),
time at clearance ("t_clear"), Cycle threshold (Ct) at peak ("c_p"),
Ct at switch ("c_s"), incubation period mean ("inc_mean"),
and incubation period standard deviation ("inc_sd").}
}
\value{
A named list including the design matrix ("design")
and a \code{data.table} of priors for covariate effects ("beta").
}
\description{
Specify the cycle threshold adjustment model formula
}
\author{
Sam Abbott
}
