% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{select_piecewise_parameters}
\alias{select_piecewise_parameters}
\title{Select the parameters to adjust in the piecewise model}
\usage{
select_piecewise_parameters(params = "all")
}
\arguments{
\item{params}{A character string indicating the paramters
to adjust in the piecewise model. Defaults to "all". Options
are: the time at peak ("t_p"), time at switch ("t_p"),
time at clearance ("t_clear"), Cycle threshold (Ct) at peak ("c_p"),
Ct at switch ("c_s"), incubation period mean ("inc_mean"),
and incubation period standard deviation ("inc_sd").}
}
\value{
A named list of parameters to adjust in the piecewise
model.
}
\description{
Select the parameters to adjust in the piecewise model
}
\examples{
select_piecewise_parameters()
}
\author{
Sam Abbott
}
