% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_variable_labels}
\alias{update_variable_labels}
\title{Update Variable Labels}
\usage{
update_variable_labels(
  draws,
  reverse = FALSE,
  variables = epict::epict_priors(include_descriptive = TRUE)
)
}
\arguments{
\item{draws}{A dataset (data.frame or data.table) containing the variables
to be relabelled.}

\item{reverse}{A logical value indicating whether to reverse the order of
variables and their corresponding labels before updating (Default is FALSE).}

\item{variables}{A data.frame providing the mapping from old variable names
to new names. By default, it uses
\code{epict::epict_priors(include_descriptive = TRUE)}. The data.frame should
contain 'variable' and 'name' columns where 'variable' is the old variable
name and 'name' is the new variable name.}
}
\value{
A data.table with updated variable labels according to the input
mapping. The returned data.table contains only the rows where variables were
found and updated.
}
\description{
This function is used to update the labels of variables in a given dataset.
The labels are updated according to a specified set of new labels.
If 'reverse' is set to TRUE, the order of variables and their corresponding
labels are reversed before the update.
}
\examples{
data <- data.frame(variable = c("v1", "v2", "v3"), value = c(1, 2, 3))
variables_mapping <- data.frame(
 variable= c("v1", "v2", "v3"),
 name = c("Variable 1", "Variable 2", "Variable 3")
)

updated_data <- update_variable_labels(data, FALSE, variables_mapping)
updated_data
}
